requirements.txt
django~=4.2.1
djangorestframework~=3.14.0
psycopg2-binary

django-filter~=23.2